[feature] check if env variable is in environ

+++ models/engine/db_storage.py
@@ -32,16 +32,17 @@ class DBStorage:
     def __init__(self):
         self.__engine = create_engine(
             'mysql+mysqldb://{}:{}@{}/{}'.format(
-                environ['HBNB_MYSQL_DB'],
                 environ['HBNB_MYSQL_USER'],
-                environ['HBNB_MYSQL_HOST'],
                 environ['HBNB_MYSQL_PWD'],
+                environ['HBNB_MYSQL_HOST'],
+                environ['HBNB_MYSQL_DB'],
                 pool_pre_ping=True))
-        if environ['HBNB_ENV'] == 'test':
-            Base.metadata.drop_all(self.__engine)
+        if "HBNB_ENV" in environ:
+            if environ['HBNB_ENV'] == 'test':
+                Base.metadata.drop_all(self.__engine)
 
     def all(self, cls=None):
-        s1 = self.session()
+        s1 = self.__session()
         store_dict = {}
         if cls is not None:
             objects = s1.query(cls).all()
@@ -50,8 +51,9 @@ class DBStorage:
                 store_dict.update({key: obj})
             return store_dict
         else:
-            for a_class in all_classes:
-                objects = s1.query(a_class).all()
+            for a_class in self.all_classes:
+                objects = s1.query(eval(a_class)).all()
+                print(objects)
                 for obj in objects:
                     key = type(a_class).__name__ + "." + obj.id
                     store_dict.update({key: obj})
@@ -59,23 +61,26 @@ class DBStorage:
 
     def new(self, obj):
         '''add object'''
-        self.session().add(obj)
-        Base.metadata.create_all(self.__engine)
+        self.__session().add(obj)
+        #Base.metadata.create_all(self.__engine)
 
     def save(self):
         '''save the session'''
-        self.session().commit()
+        self.__session().commit()
 
     def delete(self, obj=None):
         '''delete the obj'''
         if obj is not None:
-            sel = self.session.query(obj).all()
+            sel = self.__session().query(obj).all()
             for ob in sel:
-                self.session.delete(ob)
+                self.__session().delete(ob)
             self.save()
 
     def reload(self):
         '''reload database'''
-        Base.metadata.create_all(self.__engine)
+        print("reload before Base.metadata")
+        print(self.__engine)
+        #Base.metadata.create_all(self.__engine)
+        print("reload after Base.metadata")
         self.__session = sessionmaker(bind=self.__engine, expire_on_commit=False)
-        session = scoped_session(self.__session)
+        scoped_session(self.__session)
